Split ThemeJsonLoader into multiple files
