+++ AuroraEditor/Base/AppPreferences/Model/ThemePreferences/ThemeJsonLoader/ThemeJsonLoader+TextMate.swift
@@ -1,0 +1,13 @@
+//
+//  ThemeJsonLoader+TextMate.swift
+//  AuroraEditor
+//
+//  Created by TAY KAI QUAN on 16/10/22.
+//  Copyright © 2022 Aurora Company. All rights reserved.
+//
+
+import SwiftUI
+
+extension ThemeJsonLoader {
+
+}
